Move `Either` data type to newly introduced project `data-structures`
So we can have a Java 8 compatible project for common data structures.
