+++ subprojects/data-structures/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("gradlebuild.distribution.implementation-java")
+}
+
+dependencies {
+    implementation(project(":base-services"))
+}
